+++ frontend_tests/node_tests/lib/events.js
@@ -610,6 +610,7 @@ exports.fixtures = {
     typing__start: {
         type: "typing",
         op: "start",
+        message_type: "private",
         sender: typing_person1,
         recipients: [typing_person2],
     },
@@ -617,6 +618,7 @@ exports.fixtures = {
     typing__stop: {
         type: "typing",
         op: "stop",
+        message_type: "private",
         sender: typing_person1,
         recipients: [typing_person2],
     },